Initial commit.
## Rationale

The motivating reason for this is to build Swift for OpenBSD. This
effort has gotten to the point that if one wanted to, then one
could download the Swift sources, follow much of the standard build
procedures, and get a working toolchain at the end of it. But this
is not going to be normal to do for everyone and pursuing binary
packaging is the natural next step. Furthermore, it would be ideal to
run builds repeatedly and ensure that there are no build regressions.

This requires a reasonable supply of compute to achieve. One can
already use existing physical hardware to achieve this on an ongoing
basis but orchestrating this, especially across multiple platforms, is
tedious and potentially unsustainable. Moving towards a CI-style setup
means the manual toil engaged in testing against build regressions
is lessened.

Additionally, bootstrapping the Swift toolchain from scratch already
takes a number of hours. Swift has evolved to the point that one
needs an existing Swift toolchain to build Swift. Bootstrap time
can be lessened by using a prebuilt toolchain. Naturally, we want to
distribute that prebuilt toolchain and use it when building the port
for the first time, as well as having a repeatable and transparent
mechanism for sharing the build process.

There are existing compute or CI services but do not support running
OpenBSD on bare-metal or on virtual machines directly, but Linux
support is widespread. Thus using OCI containers seem like a good
fit for our purposes. GitHub additionally supports running workflows
that build container images from Containerfiles, and containers can
be stored on the GitHub Container Registry for free if they are public.

## Implementation

Ideally, we want an OpenBSD virtual machine running inside the
container, and a way to run an application within that virtual
machine without having to manually load data or input commands into
that virtual machine, as the container build process cannot handle
user input.

We need a clean operating system image. Some sources for
images exist (see https://bsd-cloud-image.org/), which come
with cloud-init. cloud-init is written in Python but has a
number of dependencies, e.g., to parse yaml. In those images, the
dependencies are preinstalled, but this causes us headaches when we
are developing our toolchain port: by virtue of a Swift toolchain
build dependency not being installed, we would get an error and adjust
our port BUILD_DEPENDS as necessary, for example. If the cloud-init
dependencies cannot be removed, then we would never know if they were
actual dependencies.

Here, we still use cloud-init, since we can use this to orchestrate
the behavior of the virtual machine, but use pyinstaller to create
a zero-dependency binary.

Finally, we need a process to install OpenBSD into the
image. Netbooting -- in the amd64 case, pxeboot -- provides the
mechanism for unattended automatic installation. This is slightly
complicated on arm64, since netbooting doesn't work correctly; the
u-boot loader can netboot the installer just fine, but the installer
doesn't detect that it has been netbooted. Unfortunately in this
case, we exploit the amd64 image to modify the ramdisk image on the
installer to make the autoinstall work without input. Maybe in the
future we can elide this step.

## Caveats

As it stands, there are probably some rough edges, specifically around
the cloud-init implementation. Some cloud-init features may not work,
but `write_files` should work properly. This is actually sufficient
for most cases, but creative uses of cloud-init may not work as yet.

KVM should probably be used and requires runner support for nested
virtualization; this thankfully is the case with GitHub workflows.

To keep the focus appropriate, the rest of the toolchain build
discussion will happen in 3405691582/swift-bootstrap, but this should
at least provide a stable base for that process and any other similar
potential applications.

narrative: true